Allow Snap API usage with redirection
In the latest API doc, Snap create transaction returns not only `token` but `redirect_url` if one wants to have a dedicated page for payment instead of JS pop-up (I do want, because I want to allow guest to put in promo code just before payment).

I can patch it in my project, but would be much better if the official version supports it, so that composer will not overwrite my patch. My proposed pull request is a quick way to return the full response while maintaining backward compatibility, but feel free to adjust the function name / design.